% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4.R
\name{Body.10}
\alias{Body.10}
\title{A Cat Function}
\usage{
Body.10(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Body.10()



}
\keyword{cats}
