% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4.R
\name{Body.26}
\alias{Body.26}
\title{A Cat Function}
\usage{
Body.26(CEPScode, ...)
}
\arguments{
\item{CEPScode, ...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{




}
\keyword{cats}
