% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EL.R
\name{EL.Huber}
\alias{EL.Huber}
\title{Empirical likelihood test for the difference of smoothed Huber estimators}
\usage{
EL.Huber(
  X,
  Y,
  mu = 0,
  conf.level = 0.95,
  scaleX = 1,
  scaleY = 1,
  VX = 2.046,
  VY = 2.046,
  k = 1.35
)
}
\arguments{
\item{X}{A numeric vector of data values.}

\item{Y}{A numeric vector of data values.}

\item{mu}{A number specifying the null hypothesis value for the difference.
Default is \code{0}.}

\item{conf.level}{Confidence level for the reported
confidence interval (default \code{0.95}).}

\item{scaleX}{The scale estimate of sample \code{X}. Default is \code{1}.}

\item{scaleY}{The scale estimate of sample \code{Y}. Default is \code{1}.}

\item{VX}{The asymptotic variance of the initial (nonsmooth) Huber estimator
for sample \code{X}. Default is \code{2.046}.}

\item{VY}{The asymptotic variance of the initial (nonsmooth) Huber estimator
for sample \code{Y}. Default is \code{2.046}.}

\item{k}{Tuning parameter for the Huber estimator. Default is \code{1.35}.}
}
\value{
A list of class \code{"htest"} with components:
\describe{
  \item{\code{estimate}}{The empirical likelihood estimate for the difference
    of two smoothed Huber estimators.}
  \item{\code{conf.int}}{A confidence interval for the difference of two
    smoothed Huber estimators.}
  \item{\code{p.value}}{The p-value for the test.}
  \item{\code{statistic}}{The value of the test statistic.}
  \item{\code{method}}{The character string
    \code{"Empirical likelihood smoothed Huber estimator difference test"}.}
  \item{\code{null.value}}{The hypothesised difference \code{mu} under \eqn{H_0}.}
  \item{\code{data.name}}{A character string giving the names of the data.}
}
}
\description{
Empirical likelihood inference for the difference of smoothed Huber estimators.
This includes a test for the null hypothesis of a constant difference of smoothed
Huber estimators, a confidence interval, and the EL estimator.
}
\details{
A common choice for a robust scale estimate (parameters \code{scaleX} and
\code{scaleY}) is the median absolute deviation (MAD).
}
\examples{
X <- rnorm(100)
Y <- rnorm(100)
t.test(X, Y)
EL.means(X, Y)
EL.Huber(X, Y)

}
\references{
Valeinis, J. and Cers, E. Extending the two-sample empirical likelihood.
Preprint: \url{http://home.lu.lv/~valeinis/lv/petnieciba/EL_TwoSample_2011.pdf}

Hampel, F., Hennig, C. and Ronchetti, E. A. (2011). A smoothing principle for
the Huber and other location M-estimators. \emph{Computational Statistics &
Data Analysis}, \strong{55}(1), 324--337.
}
\seealso{
\code{\link{EL.means}}
}
\author{
E. Cers, J. Valeinis
}
\keyword{htest}
\keyword{nonparametric}
\keyword{smooth}
