% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EL.R
\name{EL.means}
\alias{EL.means}
\title{Empirical likelihood test for the difference of two sample means}
\usage{
EL.means(X, Y, mu = 0, conf.level = 0.95)
}
\arguments{
\item{X}{A numeric vector of data values.}

\item{Y}{A numeric vector of data values.}

\item{mu}{A number specifying the null hypothesis value for the mean
difference. Default is \code{0}.}

\item{conf.level}{Confidence level for the reported
confidence interval (default \code{0.95}).}
}
\value{
A list of class \code{"htest"} with components:
\describe{
  \item{\code{estimate}}{The empirical likelihood estimate of the mean difference.}
  \item{\code{conf.int}}{A confidence interval for the mean difference.}
  \item{\code{p.value}}{The p-value for the test.}
  \item{\code{statistic}}{The value of the test statistic.}
  \item{\code{method}}{The character string \code{"Empirical likelihood mean difference test"}.}
  \item{\code{null.value}}{The hypothesised mean difference \code{mu} under \eqn{H_0}.}
  \item{\code{data.name}}{A character string giving the names of the data.}
}
}
\description{
Empirical likelihood inference for the difference of two sample means.
This includes a test for the null hypothesis of a constant mean difference,
a confidence interval, and the EL estimator.
}
\examples{
X <- rnorm(100)
Y <- rnorm(100)
t.test(X, Y)
EL.means(X, Y)
EL.Huber(X, Y)

}
\references{
Valeinis, J., Cers, E. and Cielens, J. (2010). Two-sample problems in statistical
data modelling. \emph{Mathematical Modelling and Analysis}, \strong{15}(1), 137--151.

Valeinis, J. and Cers, E. Extending the two-sample empirical likelihood.
Preprint: \url{http://home.lu.lv/~valeinis/lv/petnieciba/EL_TwoSample_2011.pdf}
}
\seealso{
\code{\link{EL.Huber}}
}
\author{
E. Cers, J. Valeinis
}
\keyword{htest}
\keyword{nonparametric}
