% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EL.R
\name{EL.statistic}
\alias{EL.statistic}
\title{The two-sample empirical likelihood statistic}
\usage{
EL.statistic(method, X, Y, Delta, d, t, bw = bw.nrd0, conf.level = 0.95)
}
\arguments{
\item{method}{One of \code{"pp"}, \code{"qq"}, \code{"roc"}, \code{"qdiff"},
or \code{"fdiff"}.}

\item{X}{A numeric vector of data values.}

\item{Y}{A numeric vector of data values.}

\item{Delta}{A number. The hypothesised value of the function at point \code{t}.}

\item{d}{Deprecated. Use \code{Delta} instead.}

\item{t}{A number. The point at which to evaluate the statistic.}

\item{bw}{A function taking a numeric vector and returning a bandwidth, or a
numeric vector of length two giving the bandwidths for \code{X} and \code{Y}
respectively. Default is \code{\link[stats]{bw.nrd0}}.}

\item{conf.level}{Confidence level for the reported
confidence interval (default \code{0.95}).}
}
\value{
An object of class \code{"htest"}.
}
\description{
Calculates \eqn{-2} times the log-likelihood ratio statistic when the function
of interest (either of P-P or Q-Q plot, ROC curve, difference of quantile or
distribution functions) at some point \code{t} is equal to \code{d}.
}
\examples{
EL.statistic(method = "pp", X = rnorm(100), Y = rnorm(100), Delta = 0.5, t = 0.5)

}
\references{
Valeinis, J. and Cers, E. Extending the two-sample empirical likelihood.
Preprint: \url{http://home.lu.lv/~valeinis/lv/petnieciba/EL_TwoSample_2011.pdf}
}
\seealso{
\code{\link{EL.smooth}}
}
\author{
E. Cers, J. Valeinis
}
\keyword{htest}
\keyword{nonparametric}
\keyword{smooth}
