% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f9.R
\name{Endpoints}
\alias{Endpoints}
\title{A Cat Function}
\usage{
Endpoints(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{




}
\keyword{cats}
