% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4.R
\name{Ep.Body}
\alias{Ep.Body}
\title{A Cat Function}
\usage{
Ep.Body(CEPScode, ...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Ep.Body()



}
\keyword{cats}
