% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2.R
\name{Ep.Header}
\alias{Ep.Header}
\title{A Cat Function}
\usage{
Ep.Header(pToken = chosenTOKEN)
}
\arguments{
\item{pToken}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{


}
\keyword{cats}
