% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3.R
\name{Ep.Query}
\alias{Ep.Query}
\title{A Cat Function}
\usage{
Ep.Query(CEPScode, ...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{



}
\keyword{cats}
