% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fet.r
\name{fei}
\alias{fei}
\title{A Cat Function}
\usage{
fei(bbbb = "3333")
}
\arguments{
\item{bbbb}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
fei()
}
\keyword{cats}
