% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{fincal}
\alias{fincal}
\title{A Cat Function}
\usage{
fincal(primarycurrency = "AUD", hoursheld = 24,
  marginavailable = 10000)
}
\arguments{
\item{primarycurrency, hoursheld, marginavailable}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{



}
\keyword{cats}
