% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{fxmath.df}
\alias{fxmath.df}
\title{A Cat Function}
\usage{
fxmath.df(Url = "https://www1.oanda.com/tools/fxcalculators/fxmath.js")
}
\arguments{
\item{Url}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
f1()

}
\keyword{cats}
