% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{GetTradeName}
\alias{GetTradeName}
\title{GetTradeName Function}
\usage{
GetTradeName(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{



}
\keyword{cats}
