% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3.R
\name{Query.6}
\alias{Query.6}
\title{A Cat Function}
\usage{
Query.6(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Query.6 ()

}
\keyword{cats}
