% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3.R
\name{Query.7}
\alias{Query.7}
\title{A Cat Function}
\usage{
Query.7(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
Query.7 ()
}
\keyword{cats}
