% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f8.R
\name{Request.EP}
\alias{Request.EP}
\title{A Cat Function}
\usage{
Request.EP(..., flat = T, REturn = "Ep.Status")
}
\arguments{
\item{..., flat, REturn}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{




}
\keyword{cats}
