% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f8.R
\name{Request.r.content}
\alias{Request.r.content}
\title{A Cat Function}
\usage{
Request.r.content(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{




}
\keyword{cats}
