% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f8.R
\name{Request.r}
\alias{Request.r}
\title{A Cat Function}
\usage{
Request.r(..., rEPScode = 1, rchosenaccountID = chosenaccountID,
  DataType = "StatusAndUrl")
}
\arguments{
\item{..., rEPScode, rchosenaccountID, DataType}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{




}
\keyword{cats}
