% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f6.R
\name{Request.Step3.30}
\alias{Request.Step3.30}
\title{A Cat Function}
\usage{
Request.Step3.30(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{



}
\keyword{cats}
