% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7.R
\name{Request.Step6.PUT}
\alias{Request.Step6.PUT}
\title{A Cat Function}
\usage{
Request.Step6.PUT(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{




}
\keyword{cats}
