% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f10.R
\name{SET.ENVIRONMENT}
\alias{SET.ENVIRONMENT}
\title{A Cat Function}
\usage{
SET.ENVIRONMENT(...)
}
\arguments{
\item{...}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{



}
\keyword{cats}
