% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{fxmath.df}
\alias{fxmath.df}
\title{fxmath.df Function}
\usage{
fxmath.df(Url = "https://www1.oanda.com/tools/fxcalculators/fxmath.js")
}
\arguments{
\item{Url}{fxmath.js url}
}
\description{
This function will GET a url and cleaned up fxmath.js into fxmath.df1 and fxmath.df2
}
\examples{
fxmath.df()

}
\keyword{fxmath.df}
